python-analyzer/requirements.txt
PyGithub>=2.1.1
matplotlib>=3.7.0
networkx>=3.2
textblob>=0.17.1
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
